Make the HttpRpc its own feature, thiserror only on std
