+++ coins/monero/src/lib.rs
@@ -1,6 +1,12 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![doc = include_str!("../README.md")]
 
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+#[macro_use]
+extern crate alloc;
+
 use std_shims::{sync::OnceLock, io};
 
 use rand_core::{RngCore, CryptoRng};
