+++ coins/monero/src/wallet/send/mod.rs
@@ -1,8 +1,6 @@
 use core::{ops::Deref, fmt};
 use std_shims::io;
 
-use thiserror::Error;
-
 use rand_core::{RngCore, CryptoRng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
 use rand::seq::SliceRandom;
@@ -42,7 +40,9 @@ use crate::{
   },
 };
 
+#[cfg(feature = "std")]
 mod builder;
+#[cfg(feature = "std")]
 pub use builder::SignableTransactionBuilder;
 
 #[cfg(feature = "multisig")]
@@ -116,34 +116,35 @@ impl SendOutput {
   }
 }
 
-#[derive(Clone, PartialEq, Eq, Debug, Error)]
+#[derive(Clone, PartialEq, Eq, Debug)]
+#[cfg_attr(feature = "std", derive(thiserror::Error))]
 pub enum TransactionError {
-  #[error("multiple addresses with payment IDs")]
+  #[cfg_attr(feature = "std", error("multiple addresses with payment IDs"))]
   MultiplePaymentIds,
-  #[error("no inputs")]
+  #[cfg_attr(feature = "std", error("no inputs"))]
   NoInputs,
-  #[error("no outputs")]
+  #[cfg_attr(feature = "std", error("no outputs"))]
   NoOutputs,
-  #[error("only one output and no change address")]
+  #[cfg_attr(feature = "std", error("only one output and no change address"))]
   NoChange,
-  #[error("too many outputs")]
+  #[cfg_attr(feature = "std", error("too many outputs"))]
   TooManyOutputs,
-  #[error("too much data")]
+  #[cfg_attr(feature = "std", error("too much data"))]
   TooMuchData,
-  #[error("too many inputs/too much arbitrary data")]
+  #[cfg_attr(feature = "std", error("too many inputs/too much arbitrary data"))]
   TooLargeTransaction,
-  #[error("not enough funds (in {0}, out {1})")]
+  #[cfg_attr(feature = "std", error("not enough funds (in {0}, out {1})"))]
   NotEnoughFunds(u64, u64),
-  #[error("wrong spend private key")]
+  #[cfg_attr(feature = "std", error("wrong spend private key"))]
   WrongPrivateKey,
-  #[error("rpc error ({0})")]
+  #[cfg_attr(feature = "std", error("rpc error ({0})"))]
   RpcError(RpcError),
-  #[error("clsag error ({0})")]
+  #[cfg_attr(feature = "std", error("clsag error ({0})"))]
   ClsagError(ClsagError),
-  #[error("invalid transaction ({0})")]
+  #[cfg_attr(feature = "std", error("invalid transaction ({0})"))]
   InvalidTransaction(RpcError),
   #[cfg(feature = "multisig")]
-  #[error("frost error {0}")]
+  #[cfg_attr(feature = "std", error("frost error {0}"))]
   FrostError(FrostError),
 }
